% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_datasets}
\alias{dm_datasets}
\title{DisMAP datasets: region and optional season}
\format{
\subsection{\code{dm_datasets}}{
}
}
\source{
\url{https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/DisMAP_Regions_CURRENT/FeatureServer/1}
}
\usage{
dm_datasets
}
\description{
A multiline simple feature boundary of regions and seasons used in DisMAP.
}
\concept{data
A data frame with 14 features and 4 fields:
- `dataset_code`: DisMAP dataset code
- `region`: Region name
- `season`: Season name
- `distribution_project_code`: DisMAP distribution project code}
\keyword{datasets}
