% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{dm_get_dataset_layers}
\alias{dm_get_dataset_layers}
\title{Get DisMAP layers (scientific species name or "Species Richness") for a dataset}
\usage{
dm_get_dataset_layers(dataset_code)
}
\arguments{
\item{datset_code}{The region (and season) code, per available \code{dataset_code} in \code{dm_datasets}}
}
\value{
A character vector of layer names, usually species scientific name or "Species Richness"
}
\description{
Get a list of available layers for a given region and season
}
\examples{
# show list of available layers for Northeast US Spring dataset
dm_get_dataset_layers("NEUS_SPR")
}
\concept{rasters}
