% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{dm_get_indicators}
\alias{dm_get_indicators}
\title{Get DisMAP indicators table, possibly filtered by field = value pairs}
\usage{
dm_get_indicators(...)
}
\arguments{
\item{...}{key = value pairs for filtering; where key could be any of the fields in the indicators table}
}
\value{
A sf object with indicators data
}
\description{
Download species indicators data including Center of Gravity metrics
}
\examples{
# Get indicators data for American Lobster in Northeast US Spring
tbl_lobster_indicators <- dm_get_indicators(
  common_name = "American lobster",
  region      = "Northeast US",
  season      = "Spring")
tbl_lobster_indicators
}
\concept{indicators}
