% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{dm_get_raster}
\alias{dm_get_raster}
\title{Get DisMAP raster for dataset layer year}
\usage{
dm_get_raster(
  dataset_code,
  layer,
  year,
  out_tif = NULL,
  out_sr = NULL,
  transform = function(x) x^(1/3),
  bbox = NULL,
  bbox_sr = NULL,
  overwrite = F,
  verbose = F
)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_datasets$dataset_code}}

\item{layer}{The layer in the dataset, typically the scientific name of the species or "Species Richness", per \code{dm_get_dataset_layers(dataset_code)}}

\item{year}{The year of the slice to download, per \code{dm_get_dataset_layer_years(dataset_code, layer)}}

\item{out_tif}{optional output GeoTIFF path to write raster}

\item{out_sr}{optional output spatial reference for output raster; defaults to native spatial reference}

\item{transform}{Function to transform raster values (default is cube root): function(x) x^(1/3)}

\item{bbox}{optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent}

\item{bbox_sr}{optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)}

\item{overwrite}{if out_tif exists, overwrite it (default is FALSE)}

\item{verbose}{if TRUE, print additional information}
}
\value{
A terra::rast object
}
\description{
Download a single interpolated biomass raster by slice ID
}
\examples{
# get raster for a specific dataset, layer, and year
dataset_code <- "NEUS_SPR"
layer <- "Paralichthys dentatus"
year <- 2015

# download the raster
r <- dm_get_raster(dataset_code, layer, year)
r

# alternatively, save to a file
tif_path <- tempfile(fileext = ".tif")
r <- dm_get_raster(dataset_code, layer, year, tif_path)
file.exists(tif_path)
}
\concept{rasters}
