% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{dm_get_survey_locations}
\alias{dm_get_survey_locations}
\title{Get DisMAP survey locations, possibly filtered by a where clause}
\usage{
dm_get_survey_locations(dataset_code, where = NULL)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_datasets$dataset_code}}

\item{where}{Optional SQL WHERE clause to filter results (e.g., "year = 2015")}
}
\value{
A sf object with survey locations
}
\description{
Download survey location points for a region
}
\examples{
# Get survey locations for Northeast US Spring 2015
dataset_code <- "NEUS_SPR"
year <- 2015

# Get locations filtered by year
pts_yr <- dm_get_survey_locations(dataset_code, where = paste0("Year = ", year))
head(pts_yr, 3)
}
\concept{surveys}
