% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{dm_plot_raster}
\alias{dm_plot_raster}
\title{Plot raster data}
\usage{
dm_plot_raster(
  r,
  interactive = TRUE,
  tiles = "Esri.OceanBasemap",
  main = NULL,
  var = NULL,
  title = NULL,
  fill_palette = ggplot2::scale_fill_viridis_c,
  ...
)
}
\arguments{
\item{r}{A SpatRaster object}

\item{interactive}{Whether to create an interactive plot (default: TRUE)}

\item{tiles}{Background tiles for interactive plot (default: "Esri.OceanBasemap")}

\item{main}{Map title for interactive plot (default: NULL)}

\item{var}{Variable name to plot in static plot (default: determined from raster)}

\item{title}{Plot title for static plot (default: NULL)}

\item{fill_palette}{Color palette function for static plot (default: scale_fill_viridis_c)}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A mapview object (interactive=TRUE) or ggplot object (interactive=FALSE)
}
\description{
Creates a visualization of the raster data, either interactively or statically.
}
\examples{
# Get a raster for demonstration
dataset_code <- "NEUS_SPR"
layer <- "Paralichthys dentatus"
year <- 2015
r <- dm_get_raster(dataset_code, layer, year)

# Interactive plot
# Not run: dm_plot_raster(r, interactive = TRUE, main = paste(dataset_code, "Summer Flounder", year))

# Static plot
dm_plot_raster(r, interactive = FALSE, title = paste(dataset_code, "Summer Flounder", year))
}
\concept{rasters}
