% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_regions}
\alias{dm_regions}
\title{DisMAP regions, seasons and dataset_code}
\format{
\subsection{\code{dm_regions}}{

A data frame with 14 features and 4 fields:
\itemize{
\item \code{dataset_code}: DisMAP dataset code
\item \code{region}: Region name
\item \code{season}: Season name
\item \code{distribution_project_code}: DisMAP distribution project code
}
}
}
\source{
\url{https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/DisMAP_Regions_CURRENT/FeatureServer/1}
}
\usage{
dm_regions
}
\description{
A multiline simple feature boundary of regions and seasons used in DisMAP.
}
\keyword{datasets}
