% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{download_dismap_raster}
\alias{download_dismap_raster}
\title{Download DisMAP Raster}
\usage{
download_dismap_raster(
  img_url,
  slice_id,
  out_tif = NULL,
  transform = function(x) x^(1/3),
  out_sr = NULL,
  bbox = NULL,
  bbox_sr = NULL,
  overwrite = F,
  verbose = F
)
}
\arguments{
\item{img_url}{The URL for the image server for a given region}

\item{slice_id}{The slice ID to download}

\item{out_tif}{optional output GeoTIFF path to write raster}

\item{transform}{Function to transform raster values (default is cube root): function(x) x^(1/3)}

\item{out_sr}{optional output spatial reference for output raster; defaults to native spatial reference}

\item{bbox}{optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent}

\item{bbox_sr}{optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)}

\item{overwrite}{if out_tif exists, overwrite it (default is FALSE)}

\item{verbose}{if TRUE, print additional information}
}
\value{
A terra::rast object
}
\description{
Download a single interpolated biomass raster by slice ID
}
