% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{get_dismap_indicators}
\alias{get_dismap_indicators}
\title{Get DisMAP Indicators Table}
\usage{
get_dismap_indicators(
  ...,
  base_url =
    "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services/Indicators_CURRENT/FeatureServer/1/query"
)
}
\arguments{
\item{base_url}{The base URL for the indicators service}

\item{common_name}{Character string of species common name}

\item{region}{Character string of desired region (optional)}
}
\value{
A sf object with indicators data
}
\description{
Download species indicators data including Center of Gravity metrics
}
