% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{get_dismap_slices}
\alias{get_dismap_slices}
\title{Get DisMAP Slice IDs}
\usage{
get_dismap_slices(species_name, region_url)
}
\arguments{
\item{species_name}{The scientific name of the species}

\item{region_url}{The URL for the image server for this region}
}
\value{
A data frame with slice information
}
\description{
Get slice IDs for a species in a region
}
