% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{get_dismap_urls}
\alias{get_dismap_urls}
\title{Get DisMAP Base URLs}
\usage{
get_dismap_urls()
}
\value{
A list with URLs for indicators, surveys, and rasters
}
\description{
Get the base URLs for various DisMAP services
}
