% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_dataset_layer_years}
\alias{get_dm_dataset_layer_years}
\title{Get DisMAP years for a dataset layer}
\usage{
get_dm_dataset_layer_years(dataset_code, layer)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_regions$dataset_code}}

\item{layer}{The layer in the dataset, typically the scientific name of the species or "Species Richness", per \code{get_dm_dataset_layers(dataset_code)}}
}
\value{
A vector of years available for the layer in the dataset
}
\description{
Get a list of years for a given species in a DisMAP dataset
}
\examples{
get_dm_dataset_layer_years("NEUS_SPR", "Paralichthys dentatus")
}
