% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_dataset_species_years}
\alias{get_dm_dataset_species_years}
\title{Get DisMAP years for a species}
\usage{
get_dm_dataset_species_years(dataset_code, species_scientific)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_regions$dataset_code}}

\item{species_scientific}{The scientific name of the species, per \code{get_dm_dataset_layers(dataset_code)}}
}
\value{
A vector of years available for the species in the dataset
}
\description{
Get a list of years for a given species in a DisMAP dataset
}
\examples{
get_dm_dataset_species_years("AI", "Paralichthys dentatus")
}
