% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_indicators}
\alias{get_dm_indicators}
\title{Get DisMAP indicators table}
\usage{
get_dm_indicators(...)
}
\arguments{
\item{...}{key = value pairs for filtering; where key could be any of the fields in the indicators table}
}
\value{
A sf object with indicators data
}
\description{
Download species indicators data including Center of Gravity metrics
}
