% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_raster}
\alias{get_dm_raster}
\title{Download DisMAP Raster}
\usage{
get_dm_raster(
  dataset_code,
  species_scientific,
  year,
  out_tif = NULL,
  out_sr = NULL,
  transform = function(x) x^(1/3),
  bbox = NULL,
  bbox_sr = NULL,
  overwrite = F,
  verbose = F
)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_regions$dataset_code}}

\item{species_scientific}{The scientific name of the species, per \code{get_dm_dataset_layers(dataset_code)}}

\item{year}{The year of the slice to download, per \code{get_dm_dataset_species_years(dataset_code, species_scientific)}}

\item{out_tif}{optional output GeoTIFF path to write raster}

\item{out_sr}{optional output spatial reference for output raster; defaults to native spatial reference}

\item{transform}{Function to transform raster values (default is cube root): function(x) x^(1/3)}

\item{bbox}{optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent}

\item{bbox_sr}{optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)}

\item{overwrite}{if out_tif exists, overwrite it (default is FALSE)}

\item{verbose}{if TRUE, print additional information}
}
\value{
A terra::rast object
}
\description{
Download a single interpolated biomass raster by slice ID
}
