% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_survey_locations}
\alias{get_dm_survey_locations}
\title{Get DisMAP survey locations}
\usage{
get_dm_survey_locations(dataset_code, where = NULL)
}
\arguments{
\item{dataset_code}{The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per \code{dm_regions$dataset_code}}

\item{where}{Optional SQL WHERE clause to filter results (e.g., "year = 2015")}
}
\value{
A sf object with survey locations
}
\description{
Download survey location points for a region
}
