% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{get_survey_locations}
\alias{get_survey_locations}
\title{Get DisMAP Survey Locations}
\usage{
get_survey_locations(
  region,
  base_url = "https://services2.arcgis.com/C8EMgrsFcRFL6LrL/ArcGIS/rest/services"
)
}
\arguments{
\item{region}{Region name}

\item{base_url}{Base URL for survey locations service}
}
\value{
A sf object with survey locations
}
\description{
Download survey location points for a region
}
