% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{list_dismap_regions}
\alias{list_dismap_regions}
\title{List Available DisMAP Regions}
\usage{
list_dismap_regions()
}
\value{
Character vector of region names
}
\description{
Return a list of available regions in DisMAP
}
