% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.r
\name{plot_dismap_distribution}
\alias{plot_dismap_distribution}
\title{Plot DisMAP Distribution Map}
\usage{
plot_dismap_distribution(
  data,
  value_col = "transformed_wtcpue",
  animated = FALSE,
  xlim = NULL,
  ylim = NULL,
  color_palette = NULL
)
}
\arguments{
\item{data}{Data frame with lon, lat, year, and transformed biomass columns}

\item{value_col}{Name of column with values to plot}

\item{animated}{Whether to create an animated plot}

\item{xlim, ylim}{Limits for x and y axes}

\item{color_palette}{Color palette function}
}
\value{
A ggplot object or animated gif
}
\description{
Create a static or animated plot of species distribution
}
