% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismapr.R
\name{get_dm_dataset_layers}
\alias{get_dm_dataset_layers}
\title{Get DisMAP layer names, given region season}
\usage{
get_dm_dataset_layers(dataset_code)
}
\arguments{
\item{datset_code}{The region (and season) code, per available \code{dataset_code} in \code{dm_regions}}
}
\value{
A character vector of layer names, usually species scientific name or "Species Richness"
}
\description{
Get a list of available layers for a given region and season
}
\examples{
get_dm_dataset_layers("AI")
}
